Copy installed files to volume
mrs produces a lot of files during installation that are required when
running. These need to be in the mrs data folder, but since we mount this
to the SSD mount point, this cannot be done while building the image.

The solution copies the generated files during image creation and copies
them back as part of the CMD that's run when an image is started in a
container.